cache gitbucket.war and plugin source tgz

+++ run_gitbucket.sh
@@ -4,9 +4,11 @@
 
 GITBUCKET_URL=https://github.com/gitbucket/gitbucket/releases/download/${GITBUCKET_VERSION}/gitbucket.war
 
-wget -q -O $HOME/gitbucket.war $GITBUCKET_URL
+if [ ! -e $HOME/Downloads/gitbucket.war ]; then
+    wget -q -O $HOME/gitbucket.war $GITBUCKET_URL
+fi
 
-nohup java -jar $HOME/gitbucket.war >$HOME/gitbucket.log 2>&1 < /dev/null &
+nohup java -jar $HOME/Downloads/gitbucket.war >$HOME/gitbucket.log 2>&1 < /dev/null &
 
 sleep 10
 cat $HOME/gitbucket.log
